Renamed tests to test HPC-5237
