+++ test/__init__.py
@@ -15,7 +15,3 @@
 """
 @author: Andy Georges (Ghent University)
 """
-from pkgutil import extend_path
-
-# we're not the only ones in this namespace
-__path__ = extend_path(__path__, __name__)  #@ReservedAssignment
